% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen.r
\name{create_api_filter}
\alias{create_api_filter}
\title{create_api_filter}
\usage{
create_api_filter(
  class_name = "ApiWrapperGenerator.HeaderFilter",
  export_modifier_pattern
)
}
\arguments{
\item{class_name}{the full class name of a CodeFileFilter.}

\item{export_modifier_pattern}{(regex?) pattern(s) to use as a marker for API functions.}
}
\value{
a CodeFileFilter
}
\description{
create_api_filter
}
