% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen.r
\name{create_rcpp_exported_func}
\alias{create_rcpp_exported_func}
\title{Creates a .NET object that parses and matches Rcpp functions declarations marked with the RcppExport attribute.}
\usage{
create_rcpp_exported_func()
}
\description{
Creates a .NET object that parses and matches Rcpp functions declarations marked with the RcppExport attribute.
}
