% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-codegen.r
\name{create_rcpp_generator_base}
\alias{create_rcpp_generator_base}
\title{Create a skeleton wrapper code gen for R package}
\usage{
create_rcpp_generator_base(prepend_header = default_rcppgen_header_prepend())
}
\arguments{
\item{prepend_header}{the text to prepend to the generated C++ code, e.g. includes and helper data marshalling functions.}
}
\description{
Create a skeleton wrapper code gen for R package
}
