% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-codegen.r
\name{create_uchronia_api_filter}
\alias{create_uchronia_api_filter}
\title{Creates a .NET object that parses and matches lines in a C API header file.}
\usage{
create_uchronia_api_filter()
}
\description{
Creates a .NET object that parses and matches lines in a C API header file.
}
