% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen.r
\name{create_wrapper_generator}
\alias{create_wrapper_generator}
\title{create_wrapper_generator}
\usage{
create_wrapper_generator(converter, api_filter)
}
\arguments{
\item{converter}{an ApiWrapperGenerator converter}

\item{api_filter}{a HeaderFilter}
}
\value{
a WrapperGenerator
}
\description{
create_wrapper_generator
}
