% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-functions.r
\name{find_env_github_root_path}
\alias{find_env_github_root_path}
\title{Find the path specified by the env var SwiftSrcPath}
\usage{
find_env_github_root_path()
}
\description{
Find the path specified by the env var SwiftSrcPath
}
