% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-functions.r
\name{find_qpp_src_path}
\alias{find_qpp_src_path}
\title{KLUDGE Find the src qpp path relative to the env var SwiftSrcPath}
\usage{
find_qpp_src_path()
}
\description{
KLUDGE Find the src qpp path relative to the env var SwiftSrcPath
}
