% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-functions.r
\name{find_uchronia_src_path}
\alias{find_uchronia_src_path}
\title{KLUDGE Find the src uchronia path relative to the env var SwiftSrcPath}
\usage{
find_uchronia_src_path()
}
\description{
KLUDGE Find the src uchronia path relative to the env var SwiftSrcPath
}
