% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-codegen.r
\name{generate_matlab_wrappers}
\alias{generate_matlab_wrappers}
\title{generate R code to wrap external pointers to domain objects.}
\usage{
generate_matlab_wrappers(
  swiftSrcPath = "",
  prepend_header = default_matlab_wrapper_prepend(),
  infile = extern_c_api_header_file_swiftr(swiftSrcPath),
  outfolder = outfolder_matlab_wrappers_swiftr(swiftSrcPath),
  api_filter = create_swift_api_filter(),
  libraryName = "mylibname"
)
}
\arguments{
\item{swiftSrcPath}{the root of swift src codebase. If empty string, find_env_SwiftSrcPath() is used.}

\item{prepend_header}{the text to prepend to the generated R code.}
}
\description{
generate R code to wrap external pointers to domain objects.
}
