% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-codegen.r
\name{generate_py_cffi_wrappers}
\alias{generate_py_cffi_wrappers}
\title{generate Python code to wrap external pointers to domain objects.}
\usage{
generate_py_cffi_wrappers(
  prepend_header = default_py_cffi_wrapper_prepend(),
  infile = extern_c_api_header_file_swiftr(swiftSrcPath),
  outfile = outfile_py_cffi_wrappers_swiftr(swiftSrcPath),
  api_filter = create_swift_api_filter(),
  cffi_obj_name = "nativelib"
)
}
\arguments{
\item{prepend_header}{the text to prepend to the generated R code.}
}
\description{
generate Python code to wrap external pointers to domain objects.
}
