% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-codegen.r
\name{generate_qpp_rcpp_glue}
\alias{generate_qpp_rcpp_glue}
\title{generate_qpp_rcpp_glue}
\usage{
generate_qpp_rcpp_glue(
  qppSrcPath,
  prepend_header = default_qpp_cppgen_prepend()
)
}
\arguments{
\item{qppSrcPath}{the root of qpp src codebase.}

\item{prepend_header}{the text to prepend to the generated C++ code, e.g. includes and helper data marshalling functions.}
}
\description{
generate C++ wrappers for the qpp R package, derived from the SWIFT C API.
}
