% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-codegen.r
\name{generate_swift_csharp_api}
\alias{generate_swift_csharp_api}
\title{generate_swift_csharp_api}
\usage{
generate_swift_csharp_api(
  swiftSrcPath = "",
  prepend_impl = default_cppgen_prepend(),
  prepend_header = default_cppgen_header_prepend()
)
}
\arguments{
\item{swiftSrcPath}{the root of swift src codebase. If empty string, find_env_SwiftSrcPath() is used.}

\item{prepend_impl}{the text to prepend to the generated C# code implementation, e.g. includes and helper data marshalling functions declarations.}
}
\description{
generate C# wrappers, derived from the SWIFT C API.
}
