% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-codegen.r
\name{generate_swift_rcpp_glue}
\alias{generate_swift_rcpp_glue}
\title{generate_swift_rcpp_glue}
\usage{
generate_swift_rcpp_glue(
  swiftSrcPath = "",
  prepend_header = default_rcppgen_header_prepend()
)
}
\arguments{
\item{prepend_header}{the text to prepend to the generated C++ code, e.g. includes and helper data marshalling functions.}

\item{`}{the root of swift src codebase. If empty string, find_env_SwiftSrcPath() is used.}
}
\description{
generate C++ wrappers for the swift R package, derived from the SWIFT C API.
}
