% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-codegen.r
\name{generate_uchronia_rcpp_glue}
\alias{generate_uchronia_rcpp_glue}
\title{generate_uchronia_rcpp_glue}
\usage{
generate_uchronia_rcpp_glue(
  uchroniaSrcPath,
  prepend_header = default_uchronia_cppgen_prepend()
)
}
\arguments{
\item{uchroniaSrcPath}{the root of uchronia src codebase.}

\item{prepend_header}{the text to prepend to the generated C++ code, e.g. includes and helper data marshalling functions.}
}
\description{
generate C++ wrappers for the uchronia R package, derived from the SWIFT C API.
}
