% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-codegen.r
\name{generate_xptr_wrappers_from_rcppfunc}
\alias{generate_xptr_wrappers_from_rcppfunc}
\title{generate R code to wrap external pointers to SWIFT objects.}
\usage{
generate_xptr_wrappers_from_rcppfunc(
  swiftSrcPath = "",
  prepend_header = default_xptr_wrapper_prepend(),
  infile = file.path(pkg_path_swiftr(swiftSrcPath), "src/rcpp_swift.cpp"),
  outfile = file.path(pkg_path_swiftr(swiftSrcPath), "R/swift-pkg-wrap-generated.r")
)
}
\arguments{
\item{swiftSrcPath}{the root of swift src codebase. If empty string, find_env_SwiftSrcPath() is used.}

\item{prepend_header}{the text to prepend to the generated R code.}
}
\description{
generate R code to wrap external pointers to SWIFT objects.
}
