% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-functions.r
\name{load_wrapper_gen_lib}
\alias{load_wrapper_gen_lib}
\title{Load the .NET assembly ApiWrapperGenerator}
\usage{
load_wrapper_gen_lib(wgenDir = "")
}
\arguments{
\item{wgenDir}{the root of swift src codebase. If empty string, find_env_SwiftSrcPath() is used to extrapolate the expected location}
}
\description{
Load the .NET assembly ApiWrapperGenerator
}
