% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-functions.r
\name{purlPkgVignettes}
\alias{purlPkgVignettes}
\title{Extract the R code from R Markdown vignettes}
\usage{
purlPkgVignettes(pkgName = "swift")
}
\arguments{
\item{pkgName}{package name}
}
\description{
Extract the R code from R Markdown vignettes
}
