% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-functions.r
\name{purlVignettes}
\alias{purlVignettes}
\title{Extract the R code from R Markdown vignettes}
\usage{
purlVignettes(outDir, pkgDir = tryFindPkgDir())
}
\arguments{
\item{outDir}{the output directory}

\item{pkgDir}{the location of the source package, where a 'vignettes' folder is expected.}
}
\description{
Extract the R code from R Markdown vignettes
}
