% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-roxygen.r
\name{roxy_joki}
\alias{roxy_joki}
\title{roxygenize joki}
\usage{
roxy_joki(swiftSrcPath = "")
}
\arguments{
\item{swiftSrcPath}{the root of swift src codebase. If empty string, find_env_SwiftSrcPath() is used.}
}
\description{
roxygenize joki
}
