% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-roxygen.r
\name{roxy_qpp}
\alias{roxy_qpp}
\title{roxygenize qpp}
\usage{
roxy_qpp(swiftSrcPath = "", pathRelativeToSwift = "../qpp/bindings/R/pkgs/qpp")
}
\arguments{
\item{swiftSrcPath}{the root of swift src codebase. If empty string, find_env_SwiftSrcPath() is used.}

\item{pathRelativeToSwift}{Location of the package to roxygenize}
}
\description{
roxygenize qpp
}
