% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-roxygen.r
\name{roxy_swiftdev}
\alias{roxy_swiftdev}
\title{roxygenize swiftdev}
\usage{
roxy_swiftdev(swiftSrcPath = "")
}
\arguments{
\item{swiftSrcPath}{the root of swift src codebase. If empty string, find_env_SwiftSrcPath() is used.}
}
\description{
roxygenize swiftdev
}
