% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-roxygen.r
\name{roxy_swiftr}
\alias{roxy_swiftr}
\title{roxygenize swift R package}
\usage{
roxy_swiftr(swiftSrcPath = "")
}
\arguments{
\item{swiftSrcPath}{the root of swift src codebase. If empty string, find_env_SwiftSrcPath() is used.}
}
\description{
roxygenize swift R package
}
