% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen.r
\name{set_wrapper_type_converter}
\alias{set_wrapper_type_converter}
\title{Defines how to create transient variables to call the wrapped API function}
\usage{
set_wrapper_type_converter(
  converter,
  api_type,
  var_postfix,
  setup_template,
  cleanup_template
)
}
\arguments{
\item{converter}{the API converter object}

\item{api_type}{the C API type, e.g. 'char**'}

\item{var_postfix}{a character to use as postfix to generate the local variable names, e.g. '_charpp'}

\item{setup_template}{C/C++ statement that specifies the creation of the local variable e.g. 'char** C_ARGNAME = createAnsiStringArray(RCPP_ARGNAME);'}

\item{cleanup_template}{C/C++ statement that specifies how to delete the local variable e.g. 'freeAnsiStringArray(C_ARGNAME, RCPP_ARGNAME.length());'}
}
\description{
Defines how to create transient variables to call the wrapped API function
}
