% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen.r
\name{set_wrapper_type_map}
\alias{set_wrapper_type_map}
\title{set_wrapper_type_map}
\usage{
set_wrapper_type_map(converter, api_type, wrapper_type)
}
\arguments{
\item{converter}{an ApiWrapperGenerator converter}

\item{api_type}{a string, the C API type, e.g. 'char**'}

\item{wrapper_type}{a string, the corresponding type in the target language, e.g. 'const std::vector<std::string>&'}
}
\description{
set_wrapper_type_map
}
