% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-functions.r
\name{swiftNativeBinDir}
\alias{swiftNativeBinDir}
\title{Gets the location of the native binaries of SWIFT}
\usage{
swiftNativeBinDir(
  srcDir = find_env_SwiftSrcPath(),
  sixtyFour = TRUE,
  buildConfig = "Debug"
)
}
\arguments{
\item{srcDir}{root of the SWIFT codebase}

\item{sixtyFour}{if TRUE look for the x64 architecture, 32 bits otherwise.}

\item{buildConfig}{The build configuration to use; 'Debug' or 'Release'}
}
\description{
Gets the location of the native binaries of SWIFT
}
