% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-codegen.r
\name{swiftr_compileAttributes}
\alias{swiftr_compileAttributes}
\title{Perform Rcpp::compileAttributes on SWIFT R}
\usage{
swiftr_compileAttributes(swiftSrcPath = "")
}
\arguments{
\item{swiftSrcPath}{the root of swift src codebase. If empty string, find_env_SwiftSrcPath() is used.}
}
\description{
Perform Rcpp::compileAttributes on SWIFT R
}
